# docs/openapi-pre-decomposition.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Stitchd Feature Flag API",
    "description": "REST API for managing feature flags, segments, and evaluation.",
    "contact": {
      "name": "Stitchd"
    },
    "license": {
      "name": "MIT",
      "identifier": "MIT"
    },
    "version": "0.1.0"
  },
  "paths": {
    "/v1/environments/{env_id}/evaluate": {
      "post": {
        "tags": [
          "flags"
        ],
        "summary": "`POST /v1/environments/{env_id}/evaluate` — Evaluate all flags for a context.",
        "description": "# Errors\nReturns [`ApiError::Database`] if evaluation logic fails.",
        "operationId": "evaluate_all_flags",
        "parameters": [
          {
            "name": "env_id",
            "in": "path",
            "description": "Environment UUID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EvaluationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Evaluation results",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BatchEvaluationResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal error"
          }
        }
      }
    },
    "/v1/environments/{env_id}/event-definitions": {
      "get": {
        "tags": [
          "event-definitions"
        ],
        "summary": "`GET /v1/environments/{env_id}/event-definitions` — List active event definitions.",
        "description": "# Errors\nReturns 500 if the repository query fails.",
        "operationId": "list_event_definitions",
        "parameters": [
          {
            "name": "env_id",
            "in": "path",
            "description": "Environment UUID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of event definitions",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EventDefinitionResponse"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal error"
          }
        }
      },
      "post": {
        "tags": [
          "event-definitions"
        ],
        "summary": "`POST /v1/environments/{env_id}/event-definitions` — Register an event definition.",
        "description": "# Errors\nReturns 409 if a definition with the same key already exists in the environment.",
        "operationId": "create_event_definition",
        "parameters": [
          {
            "name": "env_id",
            "in": "path",
            "description": "Environment UUID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateEventDefinitionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Event definition registered",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventDefinitionResponse"
                }
              }
            }
          },
          "409": {
            "description": "Key already exists in this environment"
          },
          "500": {
            "description": "Internal error"
          }
        }
      }
    },
    "/v1/environments/{env_id}/event-definitions/{key}": {
      "delete": {
        "tags": [
          "event-definitions"
        ],
        "summary": "`DELETE /v1/environments/{env_id}/event-definitions/{key}` — Soft-delete by key.",
        "description": "# Errors\nReturns 404 if no active definition with that key exists.",
        "operationId": "delete_event_definition",
        "parameters": [
          {
            "name": "env_id",
            "in": "path",
            "description": "Environment UUID",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "key",
            "in": "path",
            "description": "Event definition key",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Event definition deleted"
          },
          "404": {
            "description": "Not found"
          },
          "500": {
            "description": "Internal error"
          }
        }
      }
    },
    "/v1/environments/{env_id}/events": {
      "post": {
        "tags": [
          "events"
        ],
        "summary": "`POST /v1/environments/{env_id}/events` — Ingest a single event.",
        "description": "Validates the `metric_key` against registered definitions and checks the\nvalue type. Returns 202 immediately; the ClickHouse write is async.\n\n# Errors\n\nReturns [`ApiError`] if validation fails or an internal error occurs.",
        "operationId": "ingest_single_event",
        "parameters": [
          {
            "name": "env_id",
            "in": "path",
            "description": "Environment UUID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EventPayload"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Event accepted for ingestion"
          },
          "401": {
            "description": "Missing or invalid SDK key"
          },
          "422": {
            "description": "Unknown key or type mismatch"
          },
          "500": {
            "description": "Internal error"
          }
        },
        "security": [
          {
            "sdk_key": []
          }
        ]
      }
    },
    "/v1/environments/{env_id}/events/batch": {
      "post": {
        "tags": [
          "events"
        ],
        "summary": "`POST /v1/environments/{env_id}/events/batch` — Ingest up to 500 events.",
        "description": "All events are validated before any write is initiated. A single invalid\nevent causes the entire batch to be rejected with 422.\n\n# Errors\n\nReturns [`ApiError`] if the batch is too large, validation fails, or an internal error occurs.",
        "operationId": "ingest_batch_events",
        "parameters": [
          {
            "name": "env_id",
            "in": "path",
            "description": "Environment UUID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BatchEventPayload"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Batch accepted for ingestion"
          },
          "401": {
            "description": "Missing or invalid SDK key"
          },
          "422": {
            "description": "Batch too large, unknown key, or type mismatch"
          },
          "500": {
            "description": "Internal error"
          }
        },
        "security": [
          {
            "sdk_key": []
          }
        ]
      }
    },
    "/v1/environments/{env_id}/experiments": {
      "get": {
        "tags": [
          "experiments"
        ],
        "summary": "`GET /v1/environments/{env_id}/experiments` — List experiments in an environment.",
        "description": "# Errors\nReturns 500 on database error.",
        "operationId": "list_experiments",
        "parameters": [
          {
            "name": "env_id",
            "in": "path",
            "description": "Environment UUID",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "in": "query",
            "description": "Filter by status",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of experiments",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ExperimentResponse"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal error"
          }
        }
      },
      "post": {
        "tags": [
          "experiments"
        ],
        "summary": "`POST /v1/environments/{env_id}/experiments` — Create a new experiment.",
        "description": "# Errors\nReturns 422 if validation fails; 500 on database error.",
        "operationId": "create_experiment",
        "parameters": [
          {
            "name": "env_id",
            "in": "path",
            "description": "Environment UUID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateExperimentRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Experiment created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExperimentResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation error"
          },
          "500": {
            "description": "Internal error"
          }
        }
      }
    },
    "/v1/environments/{env_id}/experiments/{id}": {
      "get": {
        "tags": [
          "experiments"
        ],
        "summary": "`GET /v1/environments/{env_id}/experiments/{id}` — Get a single experiment.",
        "description": "# Errors\nReturns 404 if not found; 500 on database error.",
        "operationId": "get_experiment",
        "parameters": [
          {
            "name": "env_id",
            "in": "path",
            "description": "Environment UUID",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "in": "path",
            "description": "Experiment UUID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Experiment found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExperimentResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not found"
          },
          "500": {
            "description": "Internal error"
          }
        }
      },
      "delete": {
        "tags": [
          "experiments"
        ],
        "summary": "`DELETE /v1/environments/{env_id}/experiments/{id}` — Soft-delete an experiment.",
        "description": "# Errors\nReturns 404 if not found; 409 if experiment is currently running; 500 on database error.",
        "operationId": "delete_experiment",
        "parameters": [
          {
            "name": "env_id",
            "in": "path",
            "description": "Environment UUID",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "in": "path",
            "description": "Experiment UUID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Experiment deleted"
          },
          "404": {
            "description": "Not found"
          },
          "409": {
            "description": "Conflict — experiment is running"
          },
          "500": {
            "description": "Internal error"
          }
        }
      },
      "patch": {
        "tags": [
          "experiments"
        ],
        "summary": "`PATCH /v1/environments/{env_id}/experiments/{id}` — Update an experiment.",
        "description": "# Errors\nReturns 404 if not found; 409 on version conflict or invalid state; 500 on database error.",
        "operationId": "update_experiment",
        "parameters": [
          {
            "name": "env_id",
            "in": "path",
            "description": "Environment UUID",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "in": "path",
            "description": "Experiment UUID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateExperimentRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Experiment updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExperimentResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not found"
          },
          "409": {
            "description": "Conflict"
          },
          "500": {
            "description": "Internal error"
          }
        }
      }
    },
    "/v1/environments/{env_id}/experiments/{id}/iterations": {
      "get": {
        "tags": [
          "experiments"
        ],
        "summary": "`GET /v1/environments/{env_id}/experiments/{id}/iterations` — List iterations for an experiment.",
        "description": "# Errors\nReturns 404 if not found; 500 on database error.",
        "operationId": "list_iterations",
        "parameters": [
          {
            "name": "env_id",
            "in": "path",
            "description": "Environment UUID",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "in": "path",
            "description": "Experiment UUID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of iterations",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ExperimentIterationResponse"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Not found"
          },
          "500": {
            "description": "Internal error"
          }
        }
      }
    },
    "/v1/environments/{env_id}/experiments/{id}/transitions": {
      "post": {
        "tags": [
          "experiments"
        ],
        "summary": "`POST /v1/environments/{env_id}/experiments/{id}/transitions` — Transition experiment lifecycle.",
        "description": "# Errors\nReturns 404 if not found; 409 on invalid transition; 500 on database error.",
        "operationId": "transition_experiment",
        "parameters": [
          {
            "name": "env_id",
            "in": "path",
            "description": "Environment UUID",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "in": "path",
            "description": "Experiment UUID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TransitionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Experiment transitioned",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExperimentResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not found"
          },
          "409": {
            "description": "Invalid transition"
          },
          "500": {
            "description": "Internal error"
          }
        }
      }
    },
    "/v1/environments/{env_id}/segments": {
      "get": {
        "tags": [
          "segments"
        ],
        "summary": "`GET /v1/environments/{env_id}/segments` — List active segments.",
        "description": "# Errors\nReturns [`ApiError::Database`] if the repository fails to list segments.",
        "operationId": "list_segments",
        "parameters": [
          {
            "name": "env_id",
            "in": "path",
            "description": "Environment UUID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of segments",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Segment"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal error"
          }
        }
      },
      "post": {
        "tags": [
          "segments"
        ],
        "summary": "`POST /v1/environments/{env_id}/segments` — Create segment.",
        "description": "# Errors\nReturns [`ApiError::BadRequest`] if validation fails or [`ApiError::Conflict`] if the key exists.",
        "operationId": "create_segment",
        "parameters": [
          {
            "name": "env_id",
            "in": "path",
            "description": "Environment UUID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSegmentRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Segment created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Segment"
                }
              }
            }
          },
          "409": {
            "description": "Key conflict"
          },
          "422": {
            "description": "Validation error"
          }
        }
      }
    },
    "/v1/environments/{env_id}/segments/list-check": {
      "post": {
        "tags": [
          "segments"
        ],
        "summary": "`POST /v1/environments/{env_id}/segments/list-check` — Check list-segment membership.",
        "description": "Requires a valid `x-sdk-key` header for the target environment.\n\n# Errors\nReturns [`ApiError::BadRequest`] for empty segment keys, [`ApiError::Database`] on failure.",
        "operationId": "list_check_membership",
        "parameters": [
          {
            "name": "env_id",
            "in": "path",
            "description": "Environment UUID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ListCheckRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Membership results",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListCheckResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal error"
          }
        },
        "security": [
          {
            "sdk_key": []
          }
        ]
      }
    },
    "/v1/environments/{env_id}/segments/list-check/batch": {
      "post": {
        "tags": [
          "segments"
        ],
        "summary": "`POST /v1/environments/{env_id}/segments/list-check/batch` — Batch list-segment membership.",
        "description": "Requires a valid `x-sdk-key` header for the target environment.\n\n# Errors\nReturns [`ApiError::Database`] on failure.",
        "operationId": "batch_list_check_membership",
        "parameters": [
          {
            "name": "env_id",
            "in": "path",
            "description": "Environment UUID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BatchListCheckRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Batch membership results",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BatchListCheckResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal error"
          }
        },
        "security": [
          {
            "sdk_key": []
          }
        ]
      }
    },
...
    "/v1/projects/{project_id}/flags/{flag_id}/hashing": {
      "put": {
        "tags": [
          "flags"
        ],
        "summary": "`PUT /v1/projects/{project_id}/flags/{flag_id}/hashing` — Set hashing config.",
        "description": "# Errors\nReturns [`ApiError::Database`] if upsert fails.",
        "operationId": "update_hashing_config",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "description": "Project UUID",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "flag_id",
            "in": "path",
            "description": "Flag UUID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/FlagHashingConfig"
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Hashing config updated"
          },
          "500": {
            "description": "Internal error"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BatchEvaluationResponse": {
        "type": "object",
        "description": "Response containing all evaluated flags for the requested context.",
        "required": [
          "results"
        ],
        "properties": {
          "results": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/EvaluationResult"
            },
            "description": "Map of flag keys to results."
          }
        }
      },
...
      "CreateEventDefinitionRequest": {
        "type": "object",
        "description": "Request body for registering a new event definition.",
        "required": [
          "key",
          "value_type"
        ],
        "properties": {
          "key": {
            "type": "string",
            "description": "URL-safe string key, unique within the environment."
          },
          "value_type": {
            "$ref": "#/components/schemas/EventValueType",
            "description": "The metric value type this event will carry."
          }
        }
      },
...
      "EvaluationRequest": {
        "type": "object",
        "description": "Request for context-based flag evaluation.",
        "required": [
          "context"
        ],
        "properties": {
          "context": {
            "$ref": "#/components/schemas/EvaluationContext",
            "description": "Context parameters for evaluation."
          }
        }
      },
...
      "TransitionRequest": {
        "type": "object",
        "description": "Request body for transitioning an experiment to a new lifecycle status.",
        "required": [
          "to"
        ],
        "properties": {
          "to": {
            "$ref": "#/components/schemas/ExperimentStatus",
            "description": "The target lifecycle status to transition to."
          }
        }
      },
...
    }
  }
}